[WIP] Copy to NFS with subdirectories.
If you set cloud_upload_path to "nfs:<some_local_directory>" this will save the checkpoints to local disk
as normal, but then copy to this directly similar to how we do azcopy. To avoid many files in the same directly,
it adds a directory for each checkpoint. It uses these directories to discover the most recent checkpoint on restart.